+++ src/core/connect.js
@@ -1,11 +1,54 @@
+// # **connect.js**
+// **connect.js** is the facade object for accessing [jQuery's HTTP Ajax](https://api.jquery.com/jQuery.ajax/) API request.
+//
+// ## Usage
+//
+// `connect.js` is commonly setup in the `app.js` bootstrap file. To use, simply define in AMD fashion as follows:
+//
+// ```js
+// define([ 'main', ...], function (App, ...) {
+//    var connect = new Connect(AppConfig.Server);
+//    App.setConnection(connect);
+// });
+// ```
+// See an example here in the demo [demo-backbone](https://github.com/jabdul/appland/blob/demo/backbone/src/module-demo-backbone/app.js)
+//
+// From here on you can access jQuery's Ajax via the `app` instance in the files of your module.
+//
+// ```js
+// define([ 'app', ...], function (App, ...) {
+//    App.connect('json', {
+//      path: '/url/to/file.json',
+//      type: 'GET',
+//      done: function (data, textStatus) {
+//        ...
+//      },
+//      fail: function (jqXHR, textStatus, errorThrown) {
+//        throw new Error("Service ERROR: Could not fetch.");
+//      }
+//    });
+// });
+// ```
+//
+// ## Key features
+// `connect.js` provides these features:
+//
+//  - **Settings mappings**: a one-to-one mapping with the latest release of jQuery's Ajax API.
+//
+//
+// And now the API!
+
+// ## Requires
+// Include all the necessary files.
 define(['jquery'],
   function ($) {
     /**
+     * ## Connect Constructor
+     *
      * Bind to a server for AJAX requests.
      * @param {Object.<string>} o Connect's properties i.e. server's scheme,
      * port, and hostname.
      * @constructor
-     * @return {undefined}
      */
     function Connect(o) {
       /**
@@ -13,13 +56,13 @@ define(['jquery'],
        * @type {string}
        * @private
        */
-      this.protocol = o.protocol || null,
+      this.protocol = o.protocol || '';
       /**
        * Server's hostname or IP address.
        * @type {string}
        * @private
        */
-        this.hostname = o.hostname || null;
+        this.hostname = o.hostname || '';
       /**
        * Server's Port Number.
        * @type {number}
@@ -28,8 +71,7 @@ define(['jquery'],
       this.port = o.port || null;
       /**
        * The AJAX configuration and setup.
-       * @type {{data: string, path: string, cbSuccess: function(Object, string),
-       *        dataType: string, async: boolean, type: string, isLocal: string}}
+       * @type {object.<string>}
        * @private
        */
       this.requestConfig = null;
@@ -38,45 +80,55 @@ define(['jquery'],
     Connect.prototype = {
       constructor: Connect,
       /**
+       * ## Connect.prototype.request
+       *
        * Server Connection Adapter.
        * Handles the connection method requests.
        * @param {String} dataType Event's type
-       * @param {function} o The Subscriber / Observer.
-       * @return {undefined}
+       * @param {{data: string, path: string, success: function(Object, string),
+       *        dataType: string, async: boolean, type: string, isLocal: string,
+       *        cache: boolean, error: function(Object, string, *),
+       *        complete: function(Object, string),
+       *        statusCode: {}, done: function(Object, string, *),
+       *        fail: function(Object, string, *),
+       *        always: function(Object, string, *),
+       *        then: function(Object, string, *)}} o options and configuration.
        */
       request: function (dataType, o) {
         this.requestConfig = {
           data: (typeof o.data === 'undefined') ? null : o.data,
           path: o.path,
-          cbSuccess: o.cbSuccess || function (data, textStatus) {
-          },
           dataType: o.dataType || 'json',
           async: o.async || true,
           type: o.type || 'get',
           isLocal: o.isLocal || false,
           cache: o.cache || true,
-          error: o.error || function (jqXHR, textStatus, errorThrown) {
-
-            /*          console.log(o.path);
-             console.log(
-             'error(' + jqXHR + ', ' + textStatus + ', ' + errorThrown + ')'
-             );*/
-          },
-          complete: o.complete || function (jqXHR, textStatus) {
-          },
-          statusCode: o.statusCode || {
-            200: function () {
-            }
-          }
+          success: o.success ||
+            function (data, textStatus) {},
+          error: o.error ||
+            function (jqXHR, textStatus, errorThrown) {},
+          complete: o.complete ||
+            function (jqXHR, textStatus) {},
+          statusCode: o.statusCode || {},
+          done: o.done ||
+            function (data, textStatus, jqXHR) {},
+          fail: o.fail ||
+            function (data, textStatus, errorThrown) {},
+          always: o.always  ||
+            function (data, textStatus, jqXHROrErrorThrown) {},
+          then: o.then  ||
+            function (dataOrjqXHR, textStatus, jqXHROrErrorThrown) {}
         };
 
+        // Apply polymorphic operation
         if (this.requestConfig.dataType) {
           this[dataType + 'Connect']();
         }
       },
       /**
+       * ## Connect.prototype.jsonConnect
+       *
        * JSON type request.
-       * @return {undefined}
        */
       jsonConnect: function () {
         $.ajax({
@@ -88,15 +140,20 @@ define(['jquery'],
           data: this.requestConfig.data,
           cache: this.requestConfig.cache,
           async: this.requestConfig.async,
-          success: this.requestConfig.cbSuccess,
+          success: this.requestConfig.success,
           error: this.requestConfig.error,
           complete: this.requestConfig.complete,
           statusCode: this.requestConfig.statusCode
-        });
+        })
+          .done(this.requestConfig.done)
+          .fail(this.requestConfig.fail)
+          .always(this.requestConfig.always)
+          .then(this.requestConfig.then);
       },
       /**
+       * ## Connect.prototype.htmlConnect
+       *
        * HTML type request.
-       * @return {undefined}
        */
       htmlConnect: function () {
         $.ajax({
@@ -108,20 +165,28 @@ define(['jquery'],
           data: this.requestConfig.data,
           cache: this.requestConfig.cache,
           async: this.requestConfig.async,
-          success: this.requestConfig.cbSuccess,
+          success: this.requestConfig.success,
           error: this.requestConfig.error,
           complete: this.requestConfig.complete,
           statusCode: this.requestConfig.statusCode
-        });
+        })
+          .done(this.requestConfig.done)
+          .fail(this.requestConfig.fail)
+          .always(this.requestConfig.always)
+          .then(this.requestConfig.then);
       },
       /**
+       * ## Connect.prototype.getUrl
+       *
        * Create the URL from discreet segments.
        * @return {?string} The URL.
        */
       getUrl: function () {
-        var protocol = (this.protocol) ? this.protocol + '//' : 'http://',
-          hostname = this.hostname || '',
-          port = (this.port) ? ':' + this.port : '';
+        var protocol = (this.protocol) ?
+            this.protocol + '//' : 'http://',
+            hostname = this.hostname || '',
+            port = (this.port) ? ':' + this.port : '';
+
         if (hostname) {
           return (protocol + hostname + port);
         } else {