+++ src/core/util.js
@@ -1,36 +1,103 @@
-/**
- * Helper Utiltions
- *
- * Discrete functions and helpers.
- *
- * @category Util
- * @package Util
- * @author James Abdul (james.abdul@craftturf.com)
- */
+// # **util.js**
+// **util.js** is a collection of commonly used functions.
+//
+// ## Usage
+//
+// `connect.js` is accessible through `app.js` bootstrap file. To use, simply define in AMD fashion as follows:
+//
+// ```js
+// define([ 'main', ...], function (App, ...) {
+//    ...
+// });
+// ```
+// See an example here in the demo [demo-backbone](https://github.com/jabdul/appland/blob/demo/backbone/src/module-demo-backbone/app.js)
+//
+// From here on you can apply the utility functions in the files of your module as:
+//
+// ```js
+// define([ 'app', ...], function (App, ...) {
+//    if (App.Util.getDataType('stringExample') != "[object Object]") {
+//      return false;
+//    }
+//    ...
+// });
+// ```
+//
+// And now the API!
+
+// ## Requires
+// Include all the necessary files (if any).
 define(function () {
   function Util() {
+    /**
+     * # **Public API**
+     *
+     * @type {Object}
+     */
     var publicMethods = {
       /**
+       * ## extend
+       *
        * Inheritance - Prototype Inheritance
        * Inheritance using 'Parasitic combination inheritance' pattern.
        * @param {object} subType
        * @param {object} superType
+       * @return {object} subType
        */
       extend: function (subType, superType) {
-        var prototype = new Object(superType.prototype); //create prototype object
-        prototype.constructor = subType; //augment object
-        subType.prototype = prototype; //assign object
+        //create prototype object from superType
+        var prototype = new Object(superType.prototype);
+        //augment object's constructor to be that of subType
+        // i.e. it overrides properties of superType
+        prototype.constructor = subType;
+        //assign object to subType's prototype.
+        subType.prototype = prototype;
+        return subType;
       },
       /**
-       * Remove Value(s) from an Array
-       * @param {*} needle
+       * ## mixin
+       *
+       * Inheritance - Property Inheritance
+       * use of mixin approach to copy the properties of the superType
+       * to the subType.
+       * @param {object} subType
+       * @param {object} superType
+       * @return {object} subType
+       * @see http://goo.gl/IKv2eJ
+       */
+      mixin: function (subType, superType) {
+        for (var i in superType) {
+          subType[i] = superType[i];
+        }
+        return subType;
+      },
+      /**
+       * ## getDataType
+       *
+       * Get data type.
+       * This check assumes the native Object has not being overwritten by
+       * developer.
+       * @param {*} d Reference data check.
+       * @returns {string} Native constructor name if it's a reference type.
+       *                    [object Object] Native Object.
+       *                    [object Array] Native Array.
+       *                    [object Function] Native function.
+       *                    [object RegExp] Native regular expression.
+       */
+      getDataType: function (d) {
+        return Object.prototype.toString.call(d);
+      },
+      /**
+       * ## removeArrayVal
+       *
+       * Remove every occurrence of indexes with matching value.
+       * @param {*} val to test against
        * @param {Array} hayStack
-       * @param {*} replace   Default=null
        * @return {Array} hayStack
        */
-      arrayVal: function (needle, hayStack, replace) {
-        for (var i = 0; i < hayStack.length; i++) {
-          if (hayStack[i] == needle) {
+      removeArrayVal: function (val, hayStack) {
+        for (var i = 0; i < hayStack.length; i += 1) {
+          if (hayStack[i] == val) {
             hayStack.splice(i, 1);
             i--;
           }
@@ -38,13 +105,15 @@ define(function () {
         return hayStack;
       },
       /**
+       * ## inArray
+       *
        * Is value in Array?
        * @param {*} needle
        * @param {Array} hayStack
        * @return {Boolean}
        */
       inArray: function (needle, hayStack) {
-        for (var i = 0, len = hayStack.length; i < len; i++) {
+        for (var i = 0, len = hayStack.length; i < len; i += 1) {
           if (hayStack[i] === needle) {
             return true;
           }
@@ -52,12 +121,16 @@ define(function () {
         return false;
       },
       /**
-       * Fetch Query Parameters
+       * ## getQueryStringArgs
+       *
+       * Fetch Query String Parameters
        * @return {Object} Name value properties
        */
       getQueryStringArgs: function () {
         //get query string without the initial ?
-        var qs = (location.search.length > 0 ? location.search.substring(1) : ""),
+        var qs = (location.search.length > 0
+            ? location.search.substring(1)
+            : ""),
           args = {},
         //get individual items
           items = qs.length ? qs.split("&") : [],
@@ -67,7 +140,7 @@ define(function () {
           i = 0,
           len = items.length;
         //assign each item onto the args object
-        for (i; i < len; i++) {
+        for (i; i < len; i += 1) {
           item = items[i].split("=");
           name = decodeURIComponent(item[0]);
           value = decodeURIComponent(item[1]);
@@ -78,15 +151,16 @@ define(function () {
         return args;
       },
       /**
+       * ## loadStyles
+       *
        * Dynamically Load Styles
        * @param {string} url
        * @param {number} pos
-       * @return string
        */
       loadStyles: function (url, pos) {
         var link,
             head;
-        pos = (typeof pos === 'number') ? pos : 0;
+        pos = (typeof pos === 'number') ? +pos : 0;
         link = document.createElement("link");
         link.rel = "stylesheet";
         link.href = url;
@@ -94,108 +168,23 @@ define(function () {
         head.insertBefore(link, head.childNodes[pos]);
       },
       /**
+       * ## formatFloat
+       *
        * Format a Float
        * @param {string | float} value
        * @param {number} precision
        * @return string
-       * @example App.Util('123.456', 2);
-       * @see http://stackoverflow.com/questions/661562/how-to-format-a-float-in-javascript
+       * @example App.Util.formatFloat('123.456', 2);
+       * @see http://goo.gl/Udm3aV
        */
       formatFloat: function (value, precision) {
         var power = Math.pow(10, precision || 0);
-        return String((Math.round(value * power) / power).toFixed(precision));
+        return String((Math.round(value * power) / power)
+          .toFixed(precision));
       },
       /**
-       * Checks argument's type
-       * @param {string} argument The argument to test
-       * @param {string} type The variable type to test against
-       */
-      isType: function (argument, type) {
-        if (typeof argument === 'undefined') {
-          return false;
-        }
-        if (argument.constructor != type) {
-          /*console.error('Invalid argument type. Expected ' + type.name +
-           ', received ' + argument.constructor.name + ' instead.'); */
-          return false;
-        }
-        return true;
-      },
-      /**
-       * Return resource's path from given URL. Any suffix string is
-       * stripped away.
-       * @param {string} resourceUrl Resource URL.
-       * @param {string} fileExt File extension i.e jpg, gif, xls
-       * @param {boolean} withoutDomainName Excluding domain name?
-       * @param {?string} pathStartFrom Path start search from.
-       * @returns {string} Resource path or empty string.
-       */
-      getResourcePath: function (resourceUrl, fileExt, withoutDomainName, pathStartFrom) {
-        var re = new RegExp("^(.*?)\\." + '(jpg|jpeg|eps|psd|ai|tif|png)', "i"),
-          matches = re.exec(resourceUrl);
-        if (withoutDomainName) {
-          var startPos = matches[0].indexOf(pathStartFrom);
-
-          if (startPos > -1) {
-            return matches[0].substring(startPos);
-          }
-          return '';
-        }
-        return matches[0];
-      },
-      /**
-       * Returns attribute value of event's target element if exists.
-       * Works with touch devices also whereby the intended target's
-       * parent element's event may be fired instead.
-       * @param {Event} e Event object.
-       * @param {string} property For example 'className' or 'id'?
-       * @return {string}
-       */
-      getTargetAttributeValue: function (e, property) {
-        var attribute = '';
-        if (!e.target) {
-          return attribute;
-        }
-        if (e.target[property] && e.target[property] !== "") {
-          attribute = e.target[property];
-          return attribute;
-        }
-        if (e.target.parentElement && 
-          e.target.parentElement[property] !== "") {
-          attribute = e.target.parentElement[property];
-        }
-        return attribute;
-      },
-      /**
-       * Parses the UA string and extracts the version from the token.
-       * @param {boolean} tridentToken If set, returns the trident
-       *                  token value instead.
-       * @return {number} Returns the version of Internet Explorer/Trident
-       *                  or a -1.
-       * @see http://msdn.microsoft.com/en-us/library/ms537509(v=vs.85).aspx
-       */
-      getInternetExplorerVersion: function (tridentToken) {
-        var rv = -1; // Return value assumes failure
-        var ua = navigator.userAgent;
-        var matches;
-
-        if (tridentToken) {
-          var tridentRe = new RegExp("Trident/([0-9]{1,}[\.0-9]{0,})");
-          matches = tridentRe.exec(ua);
-          if (matches !== null && matches[1] !== null) {
-            rv = matches[1];
-          }
-        } else {
-          var re = new RegExp("MSIE ([0-9]{1,}[\.0-9]{0,})");
-          matches = re.exec(ua);
-          if (matches !== null && matches[1] !== null) {
-            rv = matches[1];
-          }
-        }
-
-        return rv;
-      },
-      /**
+       * ## trim
+       *
        * Trim a string
        * @param {string} s
        * @return string