+++ src/module-appland/app.js
@@ -6,11 +6,13 @@ define([
   'core/connect',
   'log4javascript'],
 function (App, Doc, AppConfig, ModuleConfig, Connect, Log4j) {
+
+  // Setup configuration for the App object.
   App.setConfig({
     'projectName': ModuleConfig['module-appland'].projectName
   });
 
-  // Setup logger
+  // Setup logger for this module
   // The getLogger string argument must be prefixed with Apl.Modules.
   ModuleConfig['module-appland'].Log =
     Log4j.getLogger( 'Apl.Modules.Appland' );
@@ -19,11 +21,15 @@ function (App, Doc, AppConfig, ModuleConfig, Connect, Log4j) {
       'setEnabled': AppConfig.App.Logging.setEnabled
   });
 
+  // Initialise this module's config.
 	App.setModuleConfig({
 		'module-appland': ModuleConfig['module-appland']
 	});
 
+  // Setup default connection to server.
+  // This can also be dynamically changed during runtime.
 	var connect = new Connect(AppConfig.Server);
 	App.setConnection(connect);
+
 	return App;
 });