+++ src/module-appland/nls/conf.js
@@ -3,11 +3,20 @@ define({
     "module-appland": {
       "projectName": 'Project Name',
       "nameSpaces":['apl-'],
-      "labels": [
-        {"LABEL_0": "WELCOME TO THE NEW ADVANCED SEARCH FROM WGSN"},
-        {"LABEL_1": "close"},
-        {"LABEL_2": "EXPLORE OVER 5M UNIQUE IMAGES"}
-      ],
+      "labels": [{
+        "LABEL_0": "Appland",
+        "LABEL_1": "Overview",
+        "LABEL_2": "Statistics overview",
+        "LABEL_3": "Modules",
+        "LABEL_4": "Active Modules",
+        "LABEL_5": "Unit Tests",
+        "LABEL_6": "Passed",
+        "LABEL_7": "Failed",
+        "LABEL_8": "Builds",
+        "LABEL_9": "Active",
+        "LABEL_10": "Module ID",
+        "LABEL_11": "This is the monitoring console for all your modules. Future versions of this admin area will contain features to help you manage your modules."
+      }],
       "pathToAssets": "../"
     }
   },